requirements.txt
# general requirements
torch==1.7.1
torchvision==0.8.2
pandas

pyyaml
tqdm
sentence_transformers